+++ Njsast/Bundler/ToplevelRenameWalker.cs
@@ -5,37 +5,33 @@ namespace Njsast.Bundler
 {
     class ToplevelRenameWalker : TreeWalker
     {
-        readonly IDictionary<string, SymbolDef> _untouchables;
-        readonly IReadOnlyDictionary<string, SymbolDef> _globals;
+        readonly IDictionary<string, SymbolDef> _variables;
+        readonly HashSet<string> _nonRootSymbolNames;
         readonly string _suffix;
 
-        public ToplevelRenameWalker(IDictionary<string, SymbolDef> untouchables,
-            IReadOnlyDictionary<string, SymbolDef> globals, string suffix)
+        public ToplevelRenameWalker(IDictionary<string, SymbolDef> variables,
+            HashSet<string> nonRootSymbolNames, string suffix)
         {
-            _untouchables = untouchables;
-            _globals = globals;
+            _variables = variables;
+            _nonRootSymbolNames = nonRootSymbolNames;
             _suffix = "_" + suffix;
         }
 
         protected override void Visit(AstNode node)
         {
-            if (!(node is AstScope scope)) return;
-            var haveToRename = scope is AstToplevel;
-            if (!haveToRename)
+            StopDescending();
+            if (!(node is AstToplevel scope))
             {
-                foreach (var symbolDef in scope.Enclosed)
-                {
-                    if (!_untouchables.ContainsKey(symbolDef.Name) && !_globals.ContainsKey(symbolDef.Name)) continue;
-                    haveToRename = true;
-                    break;
-                }
+                return;
             }
-
-            if (!haveToRename) return;
             var wasRename = false;
             foreach (var (name, symbol) in scope.Variables!)
             {
-                if (!_untouchables.ContainsKey(name) && !_globals.ContainsKey(name)) continue;
+                if (!_nonRootSymbolNames.Contains(name) && !_variables.ContainsKey(name))
+                {
+                    _variables[name] = symbol;
+                    continue;
+                }
                 string newName;
                 var index = 0;
                 do
@@ -43,14 +39,11 @@ namespace Njsast.Bundler
                     index++;
                     newName = name + _suffix;
                     if (index > 1) newName += index.ToString();
-                } while (_untouchables.ContainsKey(newName) || _globals.ContainsKey(newName));
+                } while (_nonRootSymbolNames.Contains(newName) || _variables.ContainsKey(newName));
 
                 Helpers.RenameSymbol(symbol, newName);
-                if (scope is AstToplevel)
-                {
-                    _untouchables[newName] = symbol;
-                    wasRename = true;
-                }
+                _variables[newName] = symbol;
+                wasRename = true;
             }
 
             while (wasRename)